Added repositories

+++ bookstore_backend/bookstore/src/main/java/com/listek/bookstore/models/CourierDelivery.java
@@ -2,7 +2,7 @@ package com.listek.bookstore.models;
 
 public class CourierDelivery extends Delivery{
 
-    public CourierDelivery(int id, float cost, Address address) {
+    public CourierDelivery(Long id, float cost, Address address) {
         super(id, cost, address);
     }
 }
